add writer factory, rename reader_factory to ros_factory, address review comments
