+++ src/media/CMakeLists.txt
@@ -20,8 +20,8 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/ros/ros_reader.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ros/ros_writer.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ros/ros_file_format.h"
-        "${CMAKE_CURRENT_LIST_DIR}/reader_factory.h"
-        "${CMAKE_CURRENT_LIST_DIR}/reader_factory.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/ros_factory.h"
+        "${CMAKE_CURRENT_LIST_DIR}/ros_factory.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ros_common.h"
         "${CMAKE_CURRENT_LIST_DIR}/bag_to_db3_converter.h"
         "${CMAKE_CURRENT_LIST_DIR}/bag_to_db3_converter.cpp"
