+++ src/media/ros2/ros2_writer.cpp
@@ -11,7 +11,7 @@
 #include "proc/hdr-merge.h"
 #include "proc/sequence-id-filter.h"
 #include "ros2_writer.h"
-#include "media/reader_factory.h"
+#include "media/ros_factory.h"
 #include "core/motion-frame.h"
 #include <sstream>
 #include <src/core/sensor-interface.h>
